[sign/profile]: LoginRepository로 부터 GalleryRepository 분리
