+++ data/src/main/java/org/gdsc/data/datasource/GalleryDataSource.kt
@@ -1,0 +1,5 @@
+package org.gdsc.data.datasource
+
+interface GalleryDataSource {
+    fun getGalleryImage(): ArrayList<String>
+}
