+++ data/src/main/java/org/gdsc/data/datasource/GalleryDataSourceImpl.kt
@@ -1,0 +1,45 @@
+package org.gdsc.data.datasource
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.os.Build
+import android.provider.MediaStore
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+
+private const val INDEX_MEDIA_ID = MediaStore.MediaColumns._ID
+private const val INDEX_MEDIA_URI = MediaStore.MediaColumns.DATA
+private const val INDEX_ALBUM_NAME = MediaStore.Images.Media.BUCKET_DISPLAY_NAME
+private const val INDEX_DATE_ADDED = MediaStore.MediaColumns.DATE_ADDED
+class GalleryDataSourceImpl @Inject constructor(
+    @ApplicationContext private val context: Context
+): GalleryDataSource {
+    @SuppressLint("Range")
+    override fun getGalleryImage(): ArrayList<String> {
+        val imageItemList:ArrayList<String> = ArrayList()
+        val uri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
+        val projection = arrayOf(
+            INDEX_MEDIA_ID,
+            INDEX_MEDIA_URI,
+            INDEX_ALBUM_NAME,
+            INDEX_DATE_ADDED
+        )
+
+        val selection =
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) MediaStore.Images.Media.SIZE + " > 0"
+            else null
+        val sortOrder = "$INDEX_DATE_ADDED DESC"
+        val cursor = context.contentResolver.query(uri, projection, selection, null, sortOrder)
+
+        cursor?.let {
+            while(cursor.moveToNext()) {
+                val filePath = cursor.getString(cursor.getColumnIndex(INDEX_MEDIA_URI))
+                imageItemList.add(filePath)
+            }
+        }
+
+        cursor?.close()
+
+        return imageItemList
+    }
+}
