+++ data/src/main/java/org/gdsc/data/datasource/LoginDataSource.kt
@@ -7,5 +7,4 @@ interface LoginDataSource {
 
     suspend fun postAppleToken(email: String, clientId: String): LoginResponse
 
-    fun getGalleryImage(): ArrayList<String>
 }
