+++ data/src/main/java/org/gdsc/data/datasource/LoginDataSourceImpl.kt
@@ -15,14 +15,8 @@ import java.io.File
 import javax.inject.Inject
 
 
-private const val INDEX_MEDIA_ID = MediaStore.MediaColumns._ID
-private const val INDEX_MEDIA_URI = MediaStore.MediaColumns.DATA
-private const val INDEX_ALBUM_NAME = MediaStore.Images.Media.BUCKET_DISPLAY_NAME
-private const val INDEX_DATE_ADDED = MediaStore.MediaColumns.DATE_ADDED
-
 class LoginDataSourceImpl @Inject constructor(
-    private val loginAPI: LoginAPI,
-    @ApplicationContext private val context: Context
+    private val loginAPI: LoginAPI
 ) : LoginDataSource {
     override suspend fun postGoogleToken(token: String): LoginResponse {
         return loginAPI.postUserGoogleToken(GoogleLoginRequest(token))
@@ -32,32 +26,4 @@ class LoginDataSourceImpl @Inject constructor(
         return loginAPI.postUserAppleToken(AppleLoginRequest(email, clientId))
     }
 
-    @SuppressLint("Range")
-    override fun getGalleryImage(): ArrayList<String> {
-        val imageItemList:ArrayList<String> = ArrayList()
-        val uri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-        val projection = arrayOf(
-            INDEX_MEDIA_ID,
-            INDEX_MEDIA_URI,
-            INDEX_ALBUM_NAME,
-            INDEX_DATE_ADDED
-        )
-
-        val selection =
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) MediaStore.Images.Media.SIZE + " > 0"
-            else null
-        val sortOrder = "$INDEX_DATE_ADDED DESC"
-        val cursor = context.contentResolver.query(uri, projection, selection, null, sortOrder)
-
-        cursor?.let {
-            while(cursor.moveToNext()) {
-                val filePath = cursor.getString(cursor.getColumnIndex(INDEX_MEDIA_URI))
-                imageItemList.add(filePath)
-            }
-        }
-
-        cursor?.close()
-
-        return imageItemList
-    }
 }
