+++ data/src/main/java/org/gdsc/data/di/LoginModule.kt
@@ -4,10 +4,14 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import org.gdsc.data.datasource.GalleryDataSource
+import org.gdsc.data.datasource.GalleryDataSourceImpl
 import org.gdsc.data.datasource.LoginDataSource
 import org.gdsc.data.datasource.LoginDataSourceImpl
+import org.gdsc.data.repository.GalleryRepositoryImpl
 import org.gdsc.domain.repository.LoginRepository
 import org.gdsc.data.repository.LoginRepositoryImpl
+import org.gdsc.domain.repository.GalleryRepository
 import javax.inject.Singleton
 
 @Module
@@ -21,4 +25,12 @@ abstract class LoginModule {
     @Singleton
     @Binds
     abstract fun bindLoginRepository(loginRepositoryImpl: LoginRepositoryImpl): LoginRepository
+
+    @Singleton
+    @Binds
+    abstract fun bindGalleryDataSource(galleryDataSourceImpl: GalleryDataSourceImpl): GalleryDataSource
+
+    @Singleton
+    @Binds
+    abstract fun bindGalleryRepository(galleryRepositoryImpl: GalleryRepositoryImpl): GalleryRepository
 }
