+++ data/src/main/java/org/gdsc/data/repository/GalleryRepositoryImpl.kt
@@ -1,0 +1,13 @@
+package org.gdsc.data.repository
+
+import org.gdsc.data.datasource.GalleryDataSource
+import org.gdsc.domain.repository.GalleryRepository
+import javax.inject.Inject
+
+class GalleryRepositoryImpl @Inject constructor(
+    private val galleryDataSource: GalleryDataSource
+): GalleryRepository {
+    override fun getGalleryImage(): ArrayList<String> {
+        return galleryDataSource.getGalleryImage()
+    }
+}
