+++ domain/src/main/java/org/gdsc/domain/repository/GalleryRepository.kt
@@ -1,0 +1,5 @@
+package org.gdsc.domain.repository
+
+interface GalleryRepository {
+    fun getGalleryImage(): ArrayList<String>
+}
