+++ domain/src/main/java/org/gdsc/domain/repository/LoginRepository.kt
@@ -8,5 +8,4 @@ interface LoginRepository {
 
     suspend fun postAppleToken(email: String, clientId: String): LoginResponse
 
-    fun getGalleryImage(): ArrayList<String>
 }
